# API Interface/notification-center-swagger (20-01-2025).yaml
{
  "openapi": "3.0.1",
  "info": {
    "title": "Notifacation Center API",
    "description": "API documentation for Notifacation Center service",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8091",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/api/assignments/{assignmentId}": {
      "get": {
        "tags": [
          "assignment-controller"
        ],
        "summary": "Retrieve an Assignment given its ID",
        "operationId": "getAssignmentById",
        "parameters": [
          {
            "name": "assignmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Assignment retrieved successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseAssignmentDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseMapStringString"
                }
              }
            }
          },
          "401": {
            "description": "Authentication process failed!",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          },
          "403": {
            "description": "Invalid authorization parameters. Check JWT or CSRF Token",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseMapStringString"
                }
              }
            }
          },
          "417": {
            "description": "Assignment with id [ID] not found in DB",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerToken": []
          }
        ]
      },
      "put": {
        "tags": [
          "assignment-controller"
        ],
        "summary": "Update an assignment",
        "operationId": "updateAssignment",
        "parameters": [
          {
            "name": "assignmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignmentDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Assignment updated successfully!",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseMapStringString"
                }
              }
            }
          },
          "401": {
            "description": "Authentication process failed!",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          },
          "403": {
            "description": "Invalid authorization parameters. Check JWT or CSRF Token",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseMapStringString"
                }
              }
            }
          },
          "417": {
            "description": "Expectation Failed",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          },
          "500": {
            "description": "Unable to create and store assignment",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerToken": []
          }
        ]
      },
      "delete": {
        "tags": [
          "assignment-controller"
        ],
        "summary": "Delete an assignment by ID",
        "operationId": "deleteAssignmentById_1",
        "parameters": [
          {
            "name": "assignmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Assignment deleted successfully!",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseMapStringString"
                }
              }
            }
          },
          "401": {
            "description": "Authentication process failed!",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          },
          "403": {
            "description": "Invalid authorization parameters. Check JWT or CSRF Token",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseMapStringString"
                }
              }
            }
          },
          "417": {
            "description": "Assignment with id [ID] not found in DB",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerToken": []
          }
        ]
      }
    },
...
    "/api/events": {
      "get": {
        "tags": [
          "event-controller"
        ],
        "summary": "Retrieve all events",
        "operationId": "getAllEvents",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "sortAttribute",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "timestamp"
            }
          },
          {
            "name": "isAscending",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Events retrieved successfully!",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponsePaginatedResultsDtoEventDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseMapStringString"
                }
              }
            }
          },
          "401": {
            "description": "Authentication process failed!",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          },
          "403": {
            "description": "Invalid authorization parameters. Check JWT or CSRF Token",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseMapStringString"
                }
              }
            }
          },
          "404": {
            "description": "Invalid sort attributes",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponsePaginatedResultsDtoEventDto"
                }
              }
            }
          },
          "417": {
            "description": "Expectation Failed",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerToken": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "BaseAppResponseString": {
        "type": "object",
        "properties": {
          "data": {
            "type": "string"
          },
          "errors": {
            "type": "object"
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "BaseAppResponseMapStringString": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "errors": {
            "type": "object"
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AssignmentDto": {
        "type": "object",
        "properties": {
          "assignmentId": {
            "type": "string"
          },
          "sourceUserId": {
            "type": "string"
          },
          "targetUserId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Open",
              "Accepted",
              "In Progress",
              "Completed"
            ]
          },
          "sourceUserComments": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "targetUserComments": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "description": {
            "type": "string"
          },
          "productionModule": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "timestampUpdated": {
            "type": "string",
            "format": "date-time"
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH"
            ]
          }
        }
      },
...
      "BaseAppResponsePaginatedResultsDtoEventDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PaginatedResultsDtoEventDto"
          },
          "errors": {
            "type": "object"
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "EventDto": {
        "required": [
          "pilot",
          "priority",
          "productionModule"
        ],
        "type": "object",
        "properties": {
          "eventId": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "productionModule": {
            "type": "string"
          },
          "pilot": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH"
            ]
          },
          "eventType": {
            "type": "string"
          },
          "sourceComponent": {
            "type": "string"
          },
          "smartService": {
            "type": "string"
          },
          "topic": {
            "type": "string"
          },
          "results": {
            "$ref": "#/components/schemas/JsonNode"
          }
        }
      },
      "JsonNode": {
        "type": "object"
      },
      "PaginatedResultsDtoEventDto": {
        "type": "object",
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventDto"
            }
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int32"
          },
          "lastPage": {
            "type": "boolean"
          }
        }
      },
...
      "BaseAppResponseAssignmentDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AssignmentDto"
          },
          "errors": {
            "type": "object"
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}